requirements.txt
altair==5.0.1
asynctest==0.13.0
backports.zoneinfo==0.2.1
blinker==1.4
brotlipy==0.7.0
decorator==5.1.1
einops==0.6.1
gitdb==4.0.12
GitPython==3.1.44
importlib-resources==5.12.0
Jinja2==3.1.6
jsonschema==4.17.3
kmeans-pytorch==0.3
markdown-it-py==2.2.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
opencv-python==4.8.1.78
packaging==23.2
pandas==1.3.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pyasn1-modules==0.2.8
pydeck==0.8.1b1
Pygments==2.17.2
Pympler==1.1
pyngrok==7.0.5
pyrsistent==0.19.3
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
requests-oauthlib==1.3.0
rich==13.8.1
smmap==5.0.2
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tzdata==2025.2
tzlocal==4.3.1
validators==0.20.0
watchdog==3.0.0
